Made most of the c++ examples work. Kristians still don't work, please fix

+++ doc/ref_cpp/examples/ex_query_dynamic_delete.cpp
@@ -1,5 +1,6 @@
 // @@Example: ex_cpp_dyn_query_delete @@
 #include <tightdb.hpp>
+#include <assert.h>
 
 using namespace tightdb;
 using namespace std;
@@ -38,7 +39,7 @@ int main()
 
 // @@Show@@
     // Find rows where age (column 1) < 48
-    Query query = table->where().less(1, 48);
+    Query query = table->where().less(1, (int64_t)48);
     size_t removed = query.remove();
 
     // 2 rows deleted
